+++ packages/electron-client-ipc/src/events/file.ts
@@ -1,5 +1,0 @@
-import { FileMetadata, UploadFileParams } from '../types';
-
-export interface FilesDispatchEvents {
-  createFile: (params: UploadFileParams) => { metadata: FileMetadata; success: boolean };
-}