+++ packages/electron-client-ipc/src/events/index.ts
@@ -1,9 +1,10 @@
-import { FilesDispatchEvents } from './file';
+import { LocalFilesDispatchEvents } from './localFile';
 import { MenuDispatchEvents } from './menu';
-import { FilesSearchDispatchEvents } from './search';
+import { RemoteServerBroadcastEvents, RemoteServerDispatchEvents } from './remoteServer';
 import { ShortcutDispatchEvents } from './shortcut';
 import { SystemDispatchEvents } from './system';
 import { AutoUpdateBroadcastEvents, AutoUpdateDispatchEvents } from './update';
+import { UploadFilesDispatchEvents } from './upload';
 import { WindowsDispatchEvents } from './windows';
 
 /**
@@ -12,12 +13,13 @@ import { WindowsDispatchEvents } from './windows';
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface ClientDispatchEvents
   extends WindowsDispatchEvents,
-    FilesSearchDispatchEvents,
     SystemDispatchEvents,
     MenuDispatchEvents,
-    FilesDispatchEvents,
+    LocalFilesDispatchEvents,
     AutoUpdateDispatchEvents,
-    ShortcutDispatchEvents {}
+    ShortcutDispatchEvents,
+    RemoteServerDispatchEvents,
+    UploadFilesDispatchEvents {}
 
 export type ClientDispatchEventKey = keyof ClientDispatchEvents;
 
@@ -29,7 +31,9 @@ export type ClientEventReturnType<T extends ClientDispatchEventKey> = ReturnType
  * main -> render broadcast events
  */
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface MainBroadcastEvents extends AutoUpdateBroadcastEvents {}
+export interface MainBroadcastEvents
+  extends AutoUpdateBroadcastEvents,
+    RemoteServerBroadcastEvents {}
 
 export type MainBroadcastEventKey = keyof MainBroadcastEvents;