+++ src/tools/web-browsing/index.ts
@@ -7,76 +7,70 @@ import { systemPrompt } from './systemRole';
 export const WebBrowsingApiName = {
   crawlMultiPages: 'crawlMultiPages',
   crawlSinglePage: 'crawlSinglePage',
-  searchWithSearXNG: 'searchWithSearXNG',
+  search: 'search',
 };
 
 export const WebBrowsingManifest: BuiltinToolManifest = {
   api: [
     {
       description:
-        'A meta search engine. Useful for when you need to answer questions about current events. Input should be a search query. Output is a JSON array of the query results',
-      name: WebBrowsingApiName.searchWithSearXNG,
+        'a search service. Useful for when you need to answer questions about current events. Input should be a search query. Output is a JSON array of the query results',
+      name: WebBrowsingApiName.search,
       parameters: {
         properties: {
-          optionalParams: {
-            description: 'The optional parameters for search query',
-            properties: {
-              searchCategories: {
-                description: 'The search categories you can set:',
-                items: {
-                  enum: [
-                    'files',
-                    'general',
-                    'images',
-                    'it',
-                    'map',
-                    'music',
-                    'news',
-                    'science',
-                    'social_media',
-                    'videos',
-                  ],
-                  type: 'string',
-                },
-                type: 'array',
-              },
-              searchEngines: {
-                description: 'The search engines you can use:',
-                items: {
-                  enum: [
-                    'google',
-                    'bilibili',
-                    'bing',
-                    'duckduckgo',
-                    'npm',
-                    'pypi',
-                    'github',
-                    'arxiv',
-                    'google scholar',
-                    'z-library',
-                    'reddit',
-                    'imdb',
-                    'brave',
-                    'wikipedia',
-                    'pinterest',
-                    'unsplash',
-                    'vimeo',
-                    'youtube',
-                  ],
-                  type: 'string',
-                },
-                type: 'array',
-              },
-              searchTimeRange: {
-                description: 'The time range you can set:',
-                enum: ['anytime', 'day', 'week', 'month', 'year'],
-                type: 'string',
-              },
-            },
-            type: 'object',
-          },
           query: {
             description: 'The search query',
+            type: 'string',
+          },
+          searchCategories: {
+            description: 'The search categories you can set:',
+            items: {
+              enum: [
+                'files',
+                'general',
+                'images',
+                'it',
+                'map',
+                'music',
+                'news',
+                'science',
+                'social_media',
+                'videos',
+              ],
+              type: 'string',
+            },
+            type: 'array',
+          },
+          searchEngines: {
+            description: 'The search engines you can use:',
+            items: {
+              enum: [
+                'google',
+                'bilibili',
+                'bing',
+                'duckduckgo',
+                'npm',
+                'pypi',
+                'github',
+                'arxiv',
+                'google scholar',
+                'z-library',
+                'reddit',
+                'imdb',
+                'brave',
+                'wikipedia',
+                'pinterest',
+                'unsplash',
+                'vimeo',
+                'youtube',
+              ],
+              type: 'string',
+            },
+            type: 'array',
+          },
+          searchTimeRange: {
+            description: 'The time range you can set:',
+            enum: ['anytime', 'day', 'week', 'month', 'year'],
             type: 'string',
           },
         },